% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abcd_dict}
\alias{abcd_dict}
\title{The ABCD 5.0 data dictionary}
\format{
\subsection{\code{abcd_dict}}{

A data frame with 86,599 rows and 6 columns
\describe{
\item{table_name}{Which data table the variable is from}
\item{var_name}{The variable name}
\item{var_label}{Information about the variable}
\item{notes}{Additional information about the variable}
\item{condition}{Information on whether this variable is conditioned on another}
\item{table_name_nda}{The equivalent data table name from releases 4.0 and earlier}
}
}
}
\source{
Data used in the preparation of this article were obtained from the Adolescent Brain Cognitive DevelopmentSM (ABCD) Study (https://abcdstudy.org), held in the NIMH Data Archive (NDA). This is a multisite, longitudinal study designed to recruit more than 10,000 children age 9-10 and follow them over 10 years into early adulthood. The ABCD Study® is supported by the National Institutes of Health and additional federal partners under award numbers U01DA041048, U01DA050989, U01DA051016, U01DA041022, U01DA051018, U01DA051037, U01DA050987, U01DA041174, U01DA041106, U01DA041117, U01DA041028, U01DA041134, U01DA050988, U01DA051039, U01DA041156, U01DA041025, U01DA041120, U01DA051038, U01DA041148, U01DA041093, U01DA041089, U24DA041123, U24DA041147. A full list of supporters is available at https://abcdstudy.org/federal-partners.html. A listing of participating sites and a complete listing of the study investigators can be found at https://abcdstudy.org/consortium_members/. ABCD consortium investigators designed and implemented the study and/or provided data but did not necessarily participate in the analysis or writing of this report. This manuscript reflects the views of the authors and may not reflect the opinions or views of the NIH or ABCD consortium investigators.
}
\usage{
abcd_dict
}
\description{
The ABCD 5.0 data dictionary in data.frame format
}
\keyword{datasets}
