% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{abcd_import}
\alias{abcd_import}
\title{Subset a raw ABCD dataframe by time and subjects}
\usage{
abcd_import(abcd_df, t = NULL, subjects = NULL)
}
\arguments{
\item{abcd_df}{A raw ABCD dataframe}

\item{t}{Integer representing which timepoint to filter to:
\itemize{
\item 0: baseline
\item 1: 1-year follow-up
\item 2: 2-year follow-up
\item 3: 3-year follow-up
}}

\item{subjects}{Dataframe containing list of required subjects}
}
\value{
abcd_clean_df The subsetted dataframe
}
\description{
Subset a raw ABCD dataframe by time and subjects
}
