% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{abcd_load}
\alias{abcd_load}
\title{Import an abcd csv file}
\usage{
abcd_load(csv, rename_src = TRUE)
}
\arguments{
\item{csv}{path to csv to be imported}

\item{rename_src}{if true, converts "src_subject_id" to "subjectkey"}
}
\description{
Wraps around fread and renames src_subject_id to subjectkey
}
