% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal.R
\name{abcdutils_defunct}
\alias{abcdutils_defunct}
\title{Helper function for defunct function errors.}
\usage{
abcdutils_defunct(version, alternative, env = 1)
}
\arguments{
\item{version}{Version of \code{abcdutils} in which function has been made defunct.}

\item{alternative}{Recommended alternative approach.}

\item{env}{Environment to evaluate expressions in.}
}
\value{
Returns no value. Raises an error through cli::cli_abort.
}
\description{
Helper function for defunct function errors.
}
\keyword{internal}
