% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ae}
\alias{ae}
\title{All equal function without attributes}
\usage{
ae(x, y)
}
\arguments{
\item{x}{object 1}

\item{y}{object 2}
}
\description{
Check if two objects are equal, not considering their attributes
}
