% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_scanner.R
\name{assign_site_scanner}
\alias{assign_site_scanner}
\title{Assign scanner hash IDs by randomly sampling from site}
\usage{
assign_site_scanner(
  mri_y_adm_info,
  abcd_y_lt,
  subjects = NULL,
  t = NULL,
  return_scanner_map = FALSE
)
}
\arguments{
\item{mri_y_adm_info}{Data frame storing MRI administrative information.}

\item{abcd_y_lt}{Data frame storing site information.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{return_scanner_map}{Logical indicating whether to return the scanner
map alongside the assigned scanner IDs. Defaults to \code{FALSE}.}
}
\value{
If return_scanner_map is FALSE, returns a data frame of observations
with assigned scanner IDs. Otherwise, returns a list containing that data
frame as well as the scanner map used to assign scanner IDs.
}
\description{
Some subjects in the ABCD study have no neuroimaging data, including no
scanner ID variable. Scanner effect adjustment approaches may require
scanner ID to function, but imputing a random scanner ID may be undesirable.
To avoid the need to drop observations from the imputation that do not have
scanner IDs or to impute missing scanner IDs, this function randomly
provides a scanner to subjects by sampling from the scanners available at
their site of data collection. Ensure a random seed is set prior to running
this function to obtain replicable results. Sampling of scanner ID when
multiple scanners are present at a site is weighted by the frequency of
other observations at that site with that scanner ID for that time point.
}
