% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbcl_conversion.R
\name{cbcl_raw_to_thresh}
\alias{cbcl_raw_to_thresh}
\title{Convert raw CBCL scores to thresholded ones}
\usage{
cbcl_raw_to_thresh(
  cbcl_df,
  depress_threshold_border = 5,
  depress_threshold_clinical = 7,
  anxiety_threshold_border = 6,
  anxiety_threshold_clinical = 8,
  attention_threshold_border = 9,
  attention_threshold_clinical = 12,
  aggressive_threshold_border = 11,
  aggressive_threshold_clinical = 15
)
}
\arguments{
\item{cbcl_df}{cbcl dataframe}

\item{depress_threshold_border}{cbcl threshold score}

\item{depress_threshold_clinical}{cbcl threshold score}

\item{anxiety_threshold_border}{cbcl threshold score}

\item{anxiety_threshold_clinical}{cbcl threshold score}

\item{attention_threshold_border}{cbcl threshold score}

\item{attention_threshold_clinical}{cbcl threshold score}

\item{aggressive_threshold_border}{cbcl threshold score}

\item{aggressive_threshold_clinical}{cbcl threshold score}
}
\value{
cbcl_df thresholded dataframe
}
\description{
Convert raw CBCL scores to thresholded ones
}
