% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{common_subjects}
\alias{common_subjects}
\title{Get common subjects}
\usage{
common_subjects(df_list)
}
\arguments{
\item{df_list}{List of dataframes}
}
\value{
common_subs Subjects common across list of dataframes
}
\description{
Extract subjects common across a list of dataframes
}
