% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{detail_mtbi}
\alias{detail_mtbi}
\title{Extract mTBI subjects with a minimum time-since-last-mtbi threshold}
\usage{
detail_mtbi(ph_p_otbi, abcd_y_lt, subjects = NULL, min_mpi = -10, t = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{subjects}{Dataframe containing list of required subjects}

\item{min_mpi}{The minimum time-since-last-mtbi to be selected}

\item{t}{Integer representing which timepoint to filter to:
\itemize{
\item 0: baseline
\item 1: 1-year follow-up
\item 2: 2-year follow-up
\item 3: 3-year follow-up
}}
}
\value{
subjects Dataframe containing list of required subjects
}
\description{
Extract mTBI subjects with a minimum time-since-last-mtbi threshold
}
