% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dummy}
\alias{dummy}
\title{Dummify categorical columns}
\usage{
dummy(abcd_df, cols)
}
\arguments{
\item{abcd_df}{Dataframe containing column to be dummied}

\item{cols}{Columns to be dummied}
}
\value{
abcd_df Dataframe in dummy format
}
\description{
Wrapper function for fsatDummies::dummy_cols
}
