% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{filter_subjects}
\alias{filter_subjects}
\title{Subset a dataframe to a given subject list}
\usage{
filter_subjects(abcd_df, subjects = NULL)
}
\arguments{
\item{abcd_df}{A dataframe with a subjectkey column.}

\item{subjects}{Vector of subjects to filter to.}
}
\description{
Subset a dataframe to a given subject list
}
