% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{filter_timepoint}
\alias{filter_timepoint}
\title{Subset dataframe to a collection event}
\usage{
filter_timepoint(abcd_df, t)
}
\arguments{
\item{abcd_df}{An ABCD dataframe containing a data dictionary}

\item{t}{Integer representing which timepoint to filter to:
\itemize{
\item 0: baseline
\item 1: 1-year follow-up
\item 2: 2-year follow-up
\item 3: 3-year follow-up
}}
}
\value{
timepoint_abcd_df Baseline data only
}
\description{
Subset dataframe to a collection event
}
