% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mock_data.R
\name{generate_mock_data}
\alias{generate_mock_data}
\title{Generate mock ABCD data that is safe to share.}
\usage{
generate_mock_data(df, n = NULL, seed = NULL)
}
\arguments{
\item{df}{a template ABCD dataframe}

\item{n}{number of subjects for fake dataframe. If NULL, will match the
template dataframe.}

\item{seed}{set seed before mock dataframe generation. Note that this will
alter the global environment.}
}
\value{
mock_df A scrambled abcd dataframe.
}
\description{
Given a real ABCD dataframe, generates a completely randomized counterpart
that can be used for demonstrating or practice purposes. Process shuffles
rows to ensure overall similar structure to the original ABCD dataframe in
terms of NA frequencies and within-column summary statistics.
}
