% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_all_wmnd}
\alias{get_all_wmnd}
\title{Extract white matter neurite densities}
\usage{
get_all_wmnd(...)
}
\arguments{
\item{...}{Arguments passed to \code{get_gender_y()}.}
}
\value{
Data frame containing youth-reported gender.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for extracting white matter neurite densities.
Please use \code{get_wmnd()} instead.
}
\keyword{internal}
