% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_anxiety}
\alias{get_cbcl_anxiety}
\title{Get CBCL anxiety data}
\usage{
get_cbcl_anxiety(
  mh_p_cbcl,
  subjects = NULL,
  t = NULL,
  raw = TRUE,
  anxiety_thresh_borderline = 6,
  anxiety_thresh_clinical = 8
)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{Vector of subjects to extract data for.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{raw}{Logical value indicating if raw or borderline/clinical
thresholded values should be obtained}

\item{anxiety_thresh_borderline}{threshold for borderline clinical}

\item{anxiety_thresh_clinical}{threshold for clinical}
}
\value{
cbcl_anxiety_r anxiety data
}
\description{
Get CBCL anxiety data
}
