% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_attention}
\alias{get_cbcl_attention}
\title{Get CBCL attention data}
\usage{
get_cbcl_attention(abcd_cbcls01, subjects = NULL, t = NULL, raw = TRUE)
}
\arguments{
\item{abcd_cbcls01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}

\item{raw}{Logical value indicating if raw or borderline/clinical
thresholded values should be obtained}
}
\value{
cbcl_attention_r attention data
}
\description{
Get CBCL attention data
}
