% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_depress}
\alias{get_cbcl_depress}
\title{Get CBCL depression data}
\usage{
get_cbcl_depress(
  mh_p_cbcl,
  subjects = NULL,
  t = NULL,
  raw = TRUE,
  depress_thresh_borderline = 5,
  depress_thresh_clinical = 7
)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{raw}{Logical value indicating if raw or borderline/clinical
thresholded values should be obtained}

\item{depress_thresh_borderline}{threshold for borderline clinical}

\item{depress_thresh_clinical}{threshold for clinical}
}
\value{
cbcl_depress_r depression data
}
\description{
Get CBCL depression data
}
