% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_depress_r}
\alias{get_cbcl_depress_r}
\title{Get CBCL depression data}
\usage{
get_cbcl_depress_r(abcd_cbcls01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_cbcls01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
cbcl_depress_r depression data
}
\description{
Get CBCL depression data
}
