% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_dizzy}
\alias{get_cbcl_dizzy}
\title{Get CBCL dizzy data}
\usage{
get_cbcl_dizzy(abcd_cbcl01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_cbcl01}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
cbcl_dizzy dizzy data
}
\description{
Get CBCL dizzy data
}
