% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_nausea}
\alias{get_cbcl_nausea}
\title{Get CBCL nausea data}
\usage{
get_cbcl_nausea(mh_p_cbcl, subjects = NULL, t = NULL)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
cbcl_nausea nausea data
}
\description{
Get CBCL nausea data
}
