% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_overtired}
\alias{get_cbcl_overtired}
\title{Get CBCL overtired data}
\usage{
get_cbcl_overtired(mh_p_cbcl, subjects = NULL, t = NULL)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
cbcl_overtired overtired data
}
\description{
Get CBCL overtired data
}
