% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_family_id}
\alias{get_family_id}
\title{Extract family ID}
\usage{
get_family_id(abcd_y_lt, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_y_lt}{Data frame storing family information.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing subjectkey and family ID.
}
\description{
Extract family ID
}
