% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_gender_p}
\alias{get_gender_p}
\title{return A data frame containing parent-reported genders}
\usage{
get_gender_p(gish_p_gi, subjects = NULL, t = NULL, match_y_report = FALSE)
}
\arguments{
\item{gish_p_gi}{Parent report of sex and gender data frame}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{match_y_report}{If TRUE, codes outcomes to match youth report options
at follow up timepoints (male, female, non-binary, NA). If FALSE, uses
response options provided to parents (male, female, trans male, trans
female, gender queer, and NA). In both cases, "refuse to answer" and
"don't know" are pooled with the NA option.}
}
\description{
Collected parent reported gender information from the \code{demo_gender_id_v2}
column of the \code{gish_p_gi} table.
}
