% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_gord_cor}
\alias{get_gord_cor}
\title{Extract cortical network correlations}
\usage{
get_gord_cor(mri_y_rsfmr_cor_gp_gp, subjects = NULL, t = NULL, roi_pairs)
}
\arguments{
\item{mri_y_rsfmr_cor_gp_gp}{Data file containing neurite density data}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{roi_pairs}{List where each item in the list is a 2-item vector of ROIs
to select a correlation variable for. Possible ROIs are:
- "auditory"
- "cingulo_opercular"
- "cingulo_parietal"
- "default"
- "dorsal_attention"
- "fronto_parietal"
- "none"
- "retrosplenial_temporal"
- "salience"
- "sensorimotor_hand"
- "sensorimotor_mouth"
- "ventral_attention"
- "visual"}
}
\value{
A data frame of white matter neurite densities
}
\description{
Extract cortical network correlations
}
