% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_headaches}
\alias{get_headaches}
\title{Get subject headache history}
\usage{
get_headaches(ph_p_mhx, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_mhx}{Dataframe containing medical history}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
headaches Dataframe containing headache history
}
\description{
Get subject headache history
}
