% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_income}
\alias{get_income}
\title{Returns combined household incomes split into low, medium, and high groups}
\usage{
get_income(abcd_p_demo, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_p_demo}{Dataframe containing parent demographic information}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
income_df Dataframe containing household incomes
}
\description{
Low: $0 - $50k, Medium: $50k - $100k, High: > $100k
}
