% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_loneliness}
\alias{get_loneliness}
\title{Extract loneliness}
\usage{
get_loneliness(abcd_ysr01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_ysr01}{ABCD Other Resilience}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
loneliness
}
\description{
Extract loneliness
}
