% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_age}
\alias{get_mtbi_age}
\title{Get acute symptom input variable 'latest_mtbi_age'}
\usage{
get_mtbi_age(ph_p_otbi, abcd_y_lt, subjects = NULL, t = NULL, as_years = TRUE)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{as_years}{Logical indicating whether to convert age to years}
}
\value{
mtbi_age Dataframe containing latest_mtbi_age
}
\description{
Get acute symptom input variable 'latest_mtbi_age'
}
