% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_count}
\alias{get_mtbi_count}
\title{Get number of mtbis sustained by subject}
\usage{
get_mtbi_count(otbi01, subjects = NULL, t = NULL)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
mtbi_count Dataframe containing number of previous mTBIs
}
\description{
Get number of mtbis sustained by subject
}
