% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_loc}
\alias{get_mtbi_loc}
\title{Get acute symptom input variable 'latest_mtbi_loc'}
\usage{
get_mtbi_loc(ph_p_otbi, abcd_y_lt, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI data frame}

\item{abcd_y_lt}{Data frame containing age information}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing latest_mtbi_loc
}
\description{
Get acute symptom input variable 'latest_mtbi_loc'
}
