% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_loc_l}
\alias{get_mtbi_loc_l}
\title{Get LOC data for subjects who exclusively had a latest injury in follow up}
\usage{
get_mtbi_loc_l(ph_p_otbi, abcd_y_lt, l_subs, y1_subs, y2_subs)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{l_subs}{subjects who had an mTBI exclusively in a longitudinal tp}

\item{y1_subs}{subjects who had an mTBI at year 1}

\item{y2_subs}{subjects who had an mTBI at year 2}
}
\value{
l_df latest mTBI LOC data for l subjects
}
\description{
Get LOC data for subjects who exclusively had a latest injury in follow up
}
