% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_mechanism}
\alias{get_mtbi_mechanism}
\title{Extract mechanism of latest mTBI}
\usage{
get_mtbi_mechanism(
  ph_p_otbi,
  abcd_y_lt,
  subjects = NULL,
  t = NULL,
  format = "undummied"
)
}
\arguments{
\item{ph_p_otbi}{TBI data frame}

\item{abcd_y_lt}{Data frame containing age information}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{format}{Variable indicating if df is dummied or undummied}
}
\value{
A data frame containing latest_mtbi_mechanism
}
\description{
Extract mechanism of latest mTBI
}
