% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_mem_daze_l}
\alias{get_mtbi_mem_daze_l}
\title{Get mem daze data for subjects who exclusively had a latest mtbi post bl}
\usage{
get_mtbi_mem_daze_l(abcd_lpohstbi01, l_subs, y1_subs, y2_subs)
}
\arguments{
\item{abcd_lpohstbi01}{longitudinal tbi dataframe}

\item{l_subs}{subjects who had an mTBI exclusively in a longitudinal tp}

\item{y1_subs}{subjects who had an mTBI at year 1}

\item{y2_subs}{subjects who had an mTBI at year 2}
}
\value{
l_df latest mTBI mem daze data for l subjects
}
\description{
Get mem daze data for subjects who exclusively had a latest mtbi post bl
}
