% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_nihtbx_cardsort_fc}
\alias{get_nihtbx_cardsort_fc}
\title{Get nihtbx cardsort data}
\usage{
get_nihtbx_cardsort_fc(abcd_tbss01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_tbss01}{NDA nihtbx dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
nihtbx_cardsort_fc cardsort data
}
\description{
Get nihtbx cardsort data
}
