% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_p_gender}
\alias{get_p_gender}
\title{Return dataframe containing parent-reported genders}
\usage{
get_p_gender(gish_p_gi, subjects = NULL, t = t)
}
\arguments{
\item{gish_p_gi}{Parent report of sex and gender dataframe}

\item{subjects}{Vector of subjects to extract data for.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
sex Dataframe containing sex
}
\description{
Return dataframe containing parent-reported genders
}
