% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_parent_psychopathology}
\alias{get_parent_psychopathology}
\title{Extract parent psychopathology}
\usage{
get_parent_psychopathology(mh_p_asr, subjects = NULL, t = NULL)
}
\arguments{
\item{mh_p_asr}{ABCD Parent Adult Self Report Scores Aseba}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
parent_psychopathology
}
\description{
Extract parent psychopathology
}
