% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_prosocial_behaviour}
\alias{get_prosocial_behaviour}
\title{Extract prosocial behaviour}
\usage{
get_prosocial_behaviour(psb01, abcd_psb01, subjects = NULL, t = NULL)
}
\arguments{
\item{psb01}{Parent Prosocial Behavior Survey}

\item{abcd_psb01}{Youth Prosocial Behavior Survey}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
prosocial_behaviour
}
\description{
Extract prosocial behaviour
}
