% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_pubertal_status_p}
\alias{get_pubertal_status_p}
\title{Parent report of pubertal status}
\usage{
get_pubertal_status_p(ph_p_pds, subjects = NULL, max_value = NULL, t = NULL)
}
\arguments{
\item{ph_p_pds}{Dataframe containing youth pubertal status report}

\item{subjects}{Vector of subjectkeys.}

\item{max_value}{Maximum value for pubertal status}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
pubertal_status Dataframe containing average pubertal status
}
\description{
Parent report of pubertal status
}
