% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{get_qc_data}
\alias{get_qc_data}
\title{Extract subjects that passed QC for specified neuroimaging modalities}
\usage{
get_qc_data(mri_y_qc_incl, t = NULL, subjects = NULL, type)
}
\arguments{
\item{mri_y_qc_incl}{Dataframe with QC inclusion info.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{subjects}{Vector of subjects to filter to.}

\item{type}{Which type of QC to extract data for. Valid options are:
\itemize{
\item t1w
\item t2w
\item dmri
\item rsfmri
\item mid
\item nback
\item sst
}}
}
\value{
Vector of observations that (1) passed QC, (2) failed QC, or (3) had
missing QC data.
}
\description{
Extract subjects that passed QC for specified neuroimaging modalities
}
