% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_race}
\alias{get_race}
\title{Returns race information}
\usage{
get_race(pdem02, subjects = NULL, t = t, format = "")
}
\arguments{
\item{pdem02}{Dataframe containing parent demographic information}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}

\item{format}{String indicating format to output race data}
}
\value{
race_df Dataframe containing subject race
}
\description{
Returns race information
}
