% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_scanner_id}
\alias{get_scanner_id}
\title{Extract MRI scanner serial number}
\usage{
get_scanner_id(mri_y_adm_info, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_adm_info}{Data frame storing MRI administrative information.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing subjectkey and scanner ID.
}
\description{
Extracts variable \code{mri_info_deviceserialnumber} from table \code{mri_y_adm_info}
as \code{scanner_id}.
}
