% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_screen_time}
\alias{get_screen_time}
\title{Extract healthy behaviours: screen time questionnaire}
\usage{
get_screen_time(nt_p_stq, subjects = NULL, t = NULL)
}
\arguments{
\item{nt_p_stq}{ABCD Parent Screen Time Survey}

\item{subjects}{Vector of subjects to extract data for.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
screen_time
}
\description{
Extract healthy behaviours: screen time questionnaire
}
