% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_site_id}
\alias{get_site_id}
\title{Extract site ID}
\usage{
get_site_id(abcd_y_lt, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_y_lt}{Data frame storing site information.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing subjectkey and site ID.
}
\description{
The extracted \code{site_id} variable is equal to the \code{site_id_l} variable in the
ABCD table \code{abcd_y_lt}.
}
