% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_sleep_disturbance}
\alias{get_sleep_disturbance}
\title{General factor for Sleep Disturbance from the sleep disturbance scale.}
\usage{
get_sleep_disturbance(ph_p_sds, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_sds}{ABCD table containing sleep disturbance scale data.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing sleep data.
}
\description{
This function returns a total measure for sleep disturbance from
the Sleep Disturbance Scale as defined in Bruni et al., 1996 The measure
is stored in the ABCD table \code{ph_p_sds.txt} under the variable name
\code{sds_p_ss_total}.
}
