% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{get_uninjured_subjects}
\alias{get_uninjured_subjects}
\title{Extract list of ABCD subjects who have not sustained any head injury}
\usage{
get_uninjured_subjects(abcd_otbi01, abcd_lpohstbi01)
}
\arguments{
\item{abcd_otbi01}{Baseline mtbi dataset}

\item{abcd_lpohstbi01}{Longitudinal mtbi dataset}
}
\value{
uninjured_subjects List of uninjured subjects
}
\description{
Extract list of ABCD subjects who have not sustained any head injury
}
