% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_wmnd}
\alias{get_wmnd}
\title{Extract white matter neurite densities}
\usage{
get_wmnd(drsip201, subjects = NULL, t = NULL)
}
\arguments{
\item{drsip201}{Data file containing neurite density data}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
wmnd_df Dataframe of white matter neurite densities
}
\description{
Extract white matter neurite densities
}
