% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_y_gender}
\alias{get_y_gender}
\title{Return dataframe containing youth-reported genders}
\usage{
get_y_gender(gish_y_gi, subjects = NULL, t = t)
}
\arguments{
\item{gish_y_gi}{Youth-reported gender dataframe}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
sex Dataframe containing sex
}
\description{
Return dataframe containing youth-reported genders
}
