% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{harmonize_eventname}
\alias{harmonize_eventname}
\title{Assign eventname value to a data frame}
\usage{
harmonize_eventname(df)
}
\arguments{
\item{df}{A data frame with \code{eventname} as a column.}
}
\value{
A data frame with updated eventname column.
}
\description{
Update the eventname column in a dataframe to a specific timepoint. Useful
during data cleaning for observations who are completely missing in a
particular timepoint.
}
