% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_all_tbi}
\alias{identify_all_tbi}
\title{Adds columns to tbi_df showing if a child has had an mTBI or moderate+ TBI}
\usage{
identify_all_tbi(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI data frame.}
}
\value{
A modified form of tbi_df added tbi columns.
}
\description{
Calculates whether or not child had an mTBI using the following definition:
(0 mins < LOC < 30 mins) OR (LOC < 30 mins AND felt dazed or confused)
}
\details{
Calculates moderate or severe TBIs using the following definition:
(LOC > 30 mins)

These definitions aim to match typical GCS-based definitions as well as
possible.
}
