% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_injured}
\alias{identify_injured}
\title{Identify subjects that have reported any head or neck injury}
\usage{
identify_injured(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI data frame.}
}
\description{
Identify subjects that have reported any head or neck injury
}
