% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_latest_mtbi_loc}
\alias{identify_latest_mtbi_loc}
\title{Add columns to a TBI data frame indicating the LOC of their latest mTBI}
\usage{
identify_latest_mtbi_loc(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI data frame.}
}
\value{
tbi_df with a new column indicating the LOC of the latest mTBI.
}
\description{
Add columns to a TBI data frame indicating the LOC of their latest mTBI
}
