% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_latest_mtbi_mechanism}
\alias{identify_latest_mtbi_mechanism}
\title{Add columns to a TBI data frame indicating mechanism of the latest mTBI}
\usage{
identify_latest_mtbi_mechanism(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI data frame}
}
\value{
The modified data frame
}
\description{
Add columns to a TBI data frame indicating mechanism of the latest mTBI
}
