% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{identify_mtbi_times}
\alias{identify_mtbi_times}
\title{Identify time since and age at each mTBI / most recent mTBI}
\usage{
identify_mtbi_times(tbi_df)
}
\arguments{
\item{tbi_df}{A TBI data frame}
}
\value{
The modified data frame
}
\description{
Identify time since and age at each mTBI / most recent mTBI
}
