% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{make_reader}
\alias{make_reader}
\title{Neatly read in a csv file}
\usage{
make_reader(path_maker_fn, df_it = TRUE, date = FALSE)
}
\arguments{
\item{path_maker_fn}{A function made by path_maker}

\item{df_it}{Boolean indicating if file should be read as a dataframe}

\item{date}{Boolean indicating if read file should have today (TRUE), no
date (FALSE), or a particular date (string) prefix for the file being read.
Default is FALSE - assumes the user will specify the full file name.}
}
\value{
reader function that will load a file from a particular source
}
\description{
Neatly read in a csv file
}
