% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{make_writer}
\alias{make_writer}
\title{Neatly write a csv file}
\usage{
make_writer(path_maker_fn, date = TRUE)
}
\arguments{
\item{path_maker_fn}{A function made by path_maker}

\item{date}{Boolean indicating if written file should have today (TRUE), no
date (FALSE), or a particular date (string) prefix for the file being read.
Default is TRUE - an opinionated choice to ensure that written files can't
only destroy things made on the day of the code being run.}
}
\value{
writer function that will load a file from a particular source
}
\description{
Neatly write a csv file
}
