% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numcol_to_numeric}
\alias{numcol_to_numeric}
\title{Convert dataframe columns to numeric type}
\usage{
numcol_to_numeric(df)
}
\arguments{
\item{df}{A dataframe}
}
\value{
df The dataframe with all possible columns converted to type numeric
}
\description{
Convert dataframe columns to numeric type
}
\keyword{Converts}
\keyword{a}
\keyword{all}
\keyword{be}
\keyword{can}
\keyword{columns}
\keyword{converted}
\keyword{dataframe}
\keyword{in}
\keyword{internal}
\keyword{numeric}
\keyword{that}
\keyword{to}
\keyword{type}
\keyword{type.}
