% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{open_dict}
\alias{open_dict}
\title{Open data dictionary in browser}
\usage{
open_dict()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for launching the ABCD data dictionary in browser.
Please use \code{open_dd()} instead.
}
\keyword{internal}
