% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{original_tbi_names}
\alias{original_tbi_names}
\title{Return conversion table of original and new otbi names}
\usage{
original_tbi_names()
}
\description{
Return conversion table of original and new otbi names
}
