% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{path_maker}
\alias{path_maker}
\title{Initialize a path maker}
\usage{
path_maker(path)
}
\arguments{
\item{path}{The fixed file path prefix}
}
\value{
path_fn A function that appends the fixed file path to a given name.
\code{path_fn} is also able to evaluate a \code{date} parameter. By default, \code{date}
is \code{NULL}. If a date is provided, that date will be appended as a prefix to
the provided filename. If set to \code{TRUE}, today's date (\code{Sys.Date()}) will be
used.
}
\description{
Returns a function that appends a fixed file path prefix to a given file
name.
}
\examples{
fig_path <- path_maker("/home/prashanth/figures/")
fig_path("my_plot.png")
fig_path("my_plot.png", date = "2023_03_01")
fig_path("my_plot.png", date = TRUE)

}
