% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_rsfmri}
\alias{qc_rsfmri}
\title{Extract quality control and protocol compliance information for rs-fMRI}
\usage{
qc_rsfmri(mri_y_qc_raw_rsfmr, t = 0, subjects = NULL, metric = "qc")
}
\arguments{
\item{mri_y_qc_raw_rsfmr}{Dataframe with QC info for rs-fMRI.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{subjects}{Vector of subjects to filter to.}

\item{metric}{Either "qc" (default) for quality control or "pc" for protocol
compliance.}
}
\value{
qc_results, a list of:
\enumerate{
\item Subjects that passed QC (failed by none)
\item Subjects that failed according to at least 1 rater
\item Subjects that were missing all QC data
}
}
\description{
Extract quality control and protocol compliance information for rs-fMRI
}
