% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{read_subjects}
\alias{read_subjects}
\title{Import subjectkey csv as a vector}
\usage{
read_subjects(path)
}
\arguments{
\item{path}{Path to .csv file containing subjectkey column}
}
\description{
Import subjectkey csv as a vector
}
