% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{rename_tbi}
\alias{rename_tbi}
\title{Rename non-descriptive columns in ABCD's TBI file.}
\usage{
rename_tbi(tbi_df)
}
\arguments{
\item{tbi_df}{The complete file abcd_otbi01.txt}
}
\value{
renamed_tbi A modified form of tbi_df with clearer column names
}
\description{
Renames column names of ABCD's Ohio TBI Screen data to meaningful names
that are easier to work with.
Use 'original_otbi_names()' to print out a conversion table of old and new
column names.
}
\examples{
# Mock abcd_otbi01.txt
ph_p_otbi <- data.frame(matrix(NA, nrow = 2, ncol = 107))

colnames(ph_p_otbi) <- c(
    "subjectkey",
    "eventname",
    "tbi_select_language___1",
    "tbi_1",
    "tbi_1b",
    "tbi_1c",
    "tbi_1d",
    "tbi_2",
    "tbi_2b",
    "tbi_2c",
    "tbi_2d",
    "tbi_3",
    "tbi_3b",
    "tbi_3c",
    "tbi_3d",
    "tbi_4",
    "tbi_4b",
    "tbi_4c",
    "tbi_4d",
    "tbi_5",
    "tbi_5b",
    "tbi_5c",
    "tbi_5d",
    "tbi_6o",
    "tbi_6p",
    "tbi_6q",
    "tbi_6r",
    "tbi_6s",
    "tbi_7a",
    "tbi_7c1",
    "tbl_7c2",
    "tbi_7e",
    "tbi_7f",
    "tbi_7g",
    "tbi_7i",
    "tbi_7k",
    "tbi_7l",
    "tbi_8g",
    "tbi_8i",
    "tbi_8k",
    "tbi_8l",
    "tbi_ss_ntbiloc",
    "tbi_ss_ntbiloc_nm",
    "tbi_ss_ntbiloc30",
    "tbi_ss_ntbiloc30_nm",
    "tbi_ss_nmrpi",
    "tbi_ss_nmrpi_nm",
    "tbi_ss_agefirst",
    "tbi_ss_agefirst_nm",
    "tbi_ss_before15",
    "tbi_ss_worst1",
    "tbi_ss_worst1b",
    "tbi_ss_worst2",
    "tbi_ss_worst3",
    "tbi_ss_worst4",
    "tbi_ss_worst5",
    "tbi_ss_worst_overall",
    "tbi_l_select_language___1",
    "tbi_1_l",
    "tbi_1b_l",
    "tbi_1c_l",
    "tbi_1d_l",
    "tbi_2_l",
    "tbi_2b_l",
    "tbi_2c_l",
    "tbi_2d_l",
    "tbi_3_l",
    "tbi_3b_l",
    "tbi_3c_l",
    "tbi_3d_l",
    "tbi_4_l",
    "tbi_4b_l",
    "tbi_4c_l",
    "tbi_4d_l",
    "tbi_5_l",
    "tbi_5b_l",
    "tbi_5c_l",
    "tbi_5d_l",
    "tbi_6o_l",
    "tbi_6p_l",
    "tbi_6q_l",
    "tbi_6r_l",
    "tbi_6s_l",
    "tbi_7a_l",
    "tbi_7c1_l",
    "tbl_7c2_l",
    "tbi_7e_l",
    "tbi_7f_l",
    "tbi_8",
    "tbi_ss_ntbiloc_l",
    "tbi_ss_ntbiloc_nm_l",
    "tbi_ss_ntbiloc30_l",
    "tbi_ss_ntbiloc30_nm_l",
    "tbi_ss_nmrpi_l",
    "tbi_ss_nmrpi_nm_l",
    "tbi_ss_agefirst_l",
    "tbi_ss_agefirst_nm_l",
    "tbi_ss_worst_overall_l",
    "tbi_ss_before15_nm_l",
    "tbi_ss_worst1_l",
    "tbi_ss_worst1b_l",
    "tbi_ss_worst2_l",
    "tbi_ss_worst3_l",
    "tbi_ss_worst4_l",
    "tbi_ss_worst5_l",
    "tbi_ss_agefirst_nt_l",
    "tbi_ss_before15_l"
)

ph_p_otbi_renamed <- rename_tbi(ph_p_otbi)

ph_p_otbi_renamed
}
