% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dictionary.R
\name{search_dd}
\alias{search_dd}
\title{Search the abcd data dictionary}
\usage{
search_dd(search_string)
}
\arguments{
\item{search_string}{The string to be searched}
}
\description{
Search the abcd data dictionary
}
