% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{smri_qc}
\alias{smri_qc}
\title{Remove subjects that do not pass T1w sMRI quality control}
\usage{
smri_qc(mri_y_qc_raw_smr, t, subjects = NULL, pass_method = "nofail")
}
\arguments{
\item{mri_y_qc_raw_smr}{Dataframe with QC info for T1w or T2w sMRI.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{subjects}{Vector of subjects to filter to.}

\item{pass_method}{"nofail" Assigns a pass if no rater gave a 0. "onepass"
Assigns a pass if at least one rater gave a 1. Default is "nofail".}

\item{missing_is_fail}{If TRUE (default), subjects with missing QC data are
considered to have failed QC.}
}
\value{
smri_qc_list a list of:
\enumerate{
\item Subjects that passed QC
\item Subjects that failed according to at least 1 rater
\item Subjects that failed due to missing QC data
}
}
\description{
Remove subjects that do not pass T1w sMRI quality control
}
