% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{smri_t1_qc}
\alias{smri_t1_qc}
\title{Remove subjects that do not pass T1w sMRI quality control}
\usage{
smri_t1_qc(mri_y_qc_raw_smr_t1, subjects = NULL, t = 0)
}
\arguments{
\item{mri_y_qc_raw_smr_t1}{Dataframe with QC info for T1w sMRI}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
smri_qc_list a list of:
\enumerate{
\item Subjects that passed QC
\item Subjects that failed according to at least 1 rater
\item Subjects that failed due to missing QC data
}
}
\description{
Remove subjects that do not pass T1w sMRI quality control
}
