% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_subjects}
\alias{sort_subjects}
\title{Sort a dataframe by subjectkey}
\usage{
sort_subjects(abcd_df)
}
\arguments{
\item{abcd_df}{A dataframe with a subjectkey column.}
}
\description{
Sort a dataframe by subjectkey
}
