% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{swap_src_subjectkey}
\alias{swap_src_subjectkey}
\title{Replace "src_subject_id" with "subjectkey"}
\usage{
swap_src_subjectkey(df)
}
\arguments{
\item{df}{A data frame that may have "src_subject_id" as a UID column.}
}
\value{
df with "subjectkey" as a UID column.
}
\description{
Replace "src_subject_id" with "subjectkey"
}
