% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_cbcl.R
\name{threshold_cbcl}
\alias{threshold_cbcl}
\title{Convert raw CBCL scores to thresholded ones}
\usage{
threshold_cbcl(
  cbcl_df,
  depress_thresh_border = 5,
  depress_thresh_clinical = 7,
  anxiety_thresh_border = 6,
  anxiety_thresh_clinical = 8,
  attention_thresh_border = 9,
  attention_thresh_clinical = 12,
  aggressive_thresh_border = 11,
  aggressive_thresh_clinical = 15
)
}
\arguments{
\item{cbcl_df}{cbcl dataframe}

\item{depress_thresh_border}{cbcl threshold score}

\item{depress_thresh_clinical}{cbcl threshold score}

\item{anxiety_thresh_border}{cbcl threshold score}

\item{anxiety_thresh_clinical}{cbcl threshold score}

\item{attention_thresh_border}{cbcl threshold score}

\item{attention_thresh_clinical}{cbcl threshold score}

\item{aggressive_thresh_border}{cbcl threshold score}

\item{aggressive_thresh_clinical}{cbcl threshold score}
}
\value{
cbcl_df thresholded dataframe
}
\description{
Convert raw CBCL scores to thresholded ones
}
