% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{time_subject_filter_sort}
\alias{time_subject_filter_sort}
\title{Filter dataframe by time and subjectkey, then sort by subjectkey}
\usage{
time_subject_filter_sort(abcd_df, t = NULL, subjects = NULL)
}
\arguments{
\item{abcd_df}{A dataframe with a subjectkey column.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{subjects}{Vector of subjects to filter to.}
}
\description{
Filter dataframe by time and subjectkey, then sort by subjectkey
}
