% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{vis_missing_by_df}
\alias{vis_missing_by_df}
\title{Visualize missing data on a per-dataframe basis}
\usage{
vis_missing_by_df(df_list)
}
\arguments{
\item{df_list}{Named list of dataframes}
}
\value{
missing_list list of missingness plot and missingness data
}
\description{
Visualize missing data on a per-dataframe basis
}
