% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loading_and_saving.R
\name{write_subjects}
\alias{write_subjects}
\title{Export subjectkey vector as csv}
\usage{
write_subjects(subjects, path)
}
\arguments{
\item{subjects}{Vector of subjectkeys.}

\item{path}{Path to .csv file containing subjectkey column}
}
\description{
Export subjectkey vector as csv
}
