% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_collapse}
\alias{col_collapse}
\title{Collapse two columns}
\usage{
col_collapse(df, c1, c2, new_col)
}
\arguments{
\item{df}{dataframe containing columns}

\item{c1}{string name of first column}

\item{c2}{string name of second column}

\item{new_col}{string name of collapsed column}
}
\value{
RETURN
}
\description{
When two columns contain complementary information (i.e., when one column has
a value, the other must always be NA), replace both columns with a single
one.
}
