% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{detail_mtbi}
\alias{detail_mtbi}
\title{Chain several TBI annotation functions to add multiple helpful mTBI columns}
\usage{
detail_mtbi(otbi01, subjects, t = NULL)
}
\arguments{
\item{otbi01}{The baseline TBI dataframe}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{Integer representing which timepoint to filter to:
\itemize{
\item 0: baseline
\item 1: 1-year follow-up
\item 2: 2-year follow-up
\item 3: 3-year follow-up
}}
}
\value{
detailed_otbi01 The modified dataframe
}
\description{
Chain several TBI annotation functions to add multiple helpful mTBI columns
}
