% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetting.R
\name{filter_timepoint}
\alias{filter_timepoint}
\title{Subset dataframe to a collection event}
\usage{
filter_timepoint(abcd_df, t)
}
\arguments{
\item{abcd_df}{An ABCD dataframe.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\description{
Subset dataframe to a collection event
}
