% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{format_race}
\alias{format_race}
\title{Print summary of subjects by race}
\usage{
format_race(race_df, format)
}
\arguments{
\item{race_df}{Data frame generated by race extraction function}

\item{format}{Variable indicating if df is dummied or undummied}
}
\value{
summary table of subjects by race
}
\description{
Print summary of subjects by race
}
