% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_all_wmnd}
\alias{get_all_wmnd}
\title{Extract white matter neurite densities}
\usage{
get_all_wmnd(mri_y_rsi_rnd_at, mri_y_rsi_rnd_wm_dsk, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_rsi_rnd_at}{Data file containing neurite density data}

\item{mri_y_rsi_rnd_wm_dsk}{Data file containing neurite density data}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame of white matter neurite densities
}
\description{
Extract neurite densities for major white matter tracts (AtlasTrack ROIs) as
well as peri-cortical/sub-adjacent white matter structures defined relative
to the Desikan Cortical Parcellation.
}
