% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_attention}
\alias{get_cbcl_attention}
\title{Get CBCL attention data}
\usage{
get_cbcl_attention(
  mh_p_cbcl,
  subjects = NULL,
  t = NULL,
  raw = TRUE,
  attention_thresh_borderline = 9,
  attention_thresh_clinical = 12
)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{raw}{Logical value indicating if raw or borderline/clinical
thresholded values should be obtained}

\item{attention_thresh_borderline}{threshold for borderline clinical}

\item{attention_thresh_clinical}{threshold for clinical}
}
\value{
cbcl_attention_r attention data
}
\description{
Get CBCL attention data
}
