% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_sleeping_more}
\alias{get_cbcl_sleeping_more}
\title{Get CBCL sleeping_more data}
\usage{
get_cbcl_sleeping_more(mh_p_cbcl, subjects = NULL, t = NULL)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{list of subjects to receive data for}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
cbcl_sleeping_more sleeping_more data
}
\description{
Get CBCL sleeping_more data
}
