% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_syndrome_scale}
\alias{get_cbcl_syndrome_scale}
\title{Extract CBCL syndrome scale data}
\usage{
get_cbcl_syndrome_scale(
  mh_p_cbcl,
  syndrome,
  raw = TRUE,
  t = NULL,
  subjects = NULL
)
}
\arguments{
\item{mh_p_cbcl}{Dataframe containing ABCD CBCL data}

\item{syndrome}{String indicating which of the 8 CBCL syndrome scales should
be collected. Options include: "anxdep", "withdep", "somatic", "social",
"thought", "attention", "rulebreak", and "aggressive"}

\item{raw}{Boolean indicating if extracted data should be raw (TRUE) or
t-scores (FALSE). Defaults to TRUE.}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}

\item{subjects}{list of subjects to receive data for}
}
\value{
ss_data Dataframe containing syndrome scale data
}
\description{
Extract CBCL syndrome scale data
}
