% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cbcl_vomiting}
\alias{get_cbcl_vomiting}
\title{Get CBCL vomiting data}
\usage{
get_cbcl_vomiting(mh_p_cbcl, subjects = NULL, t = NULL)
}
\arguments{
\item{mh_p_cbcl}{NDA cbcl dataframe}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
cbcl_vomiting vomiting data
}
\description{
Get CBCL vomiting data
}
