% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cort_sa}
\alias{get_cort_sa}
\title{Extract cortical surface areas}
\usage{
get_cort_sa(mri_y_smr_area_dsk, subjects = NULL, t = NULL, only_total = TRUE)
}
\arguments{
\item{mri_y_smr_area_dsk}{Data file containing cortical data}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{only_total}{Logical indicating whether to return only the total
cortical surface area or all cortical surface areas. Defaults to TRUE}
}
\value{
A data frame of cortical surface areas
}
\description{
Extract cortical surface areas
}
