% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_cort_t}
\alias{get_cort_t}
\title{Extract cortical thicknesses}
\usage{
get_cort_t(mri_y_smr_thk_dsk, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_smr_thk_dsk}{Data file containing cortical data}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
cort_t_df Dataframe of cortical thicknesses
}
\description{
Extract cortical thicknesses
}
