% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_exercise}
\alias{get_exercise}
\title{Extract healthy behaviours: exercise questionnaire}
\usage{
get_exercise(abcd_yrb01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_yrb01}{ABCD Youth Youth Risk Behavior Survey Exercise Physical
Activity}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
exercise
}
\description{
Extract healthy behaviours: exercise questionnaire
}
