% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_family_function}
\alias{get_family_function}
\title{Extract family function}
\usage{
get_family_function(ce_y_fes, ce_p_fes, subjects = NULL, t = NULL)
}
\arguments{
\item{ce_y_fes}{ABCD Parent Family Environment Scale-Family Conflict Subscale
Modified from PhenX}

\item{ce_p_fes}{ABCD Parent Family Environment Scale-Family Conflict
Subscale Modified from PhenX}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
family_function
}
\description{
Extract family function
}
