% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_friends}
\alias{get_friends}
\title{Extract number of friends}
\usage{
get_friends(mh_y_or, gish_p_gi, subjects = NULL, discretize = FALSE, t = NULL)
}
\arguments{
\item{mh_y_or}{ABCD Other Resilience}

\item{gish_p_gi}{Parent report of sex and gender dataframe}

\item{subjects}{Vector of subjectkeys.}

\item{discretize}{Boolean indicating if the loneliness measures should be
discretized into quartiles}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\description{
Extract number of friends
}
