% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_gord_cor}
\alias{get_gord_cor}
\title{Extract cortical network correlations}
\usage{
get_gord_cor(betnet02, subjects = NULL, t = NULL)
}
\arguments{
\item{betnet02}{Data file containing neurite density data}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
gord_cor Dataframe of white matter neurite densities
}
\description{
Extract cortical network correlations
}
