% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_gord_cor}
\alias{get_gord_cor}
\title{Extract cortical network correlations}
\usage{
get_gord_cor(mri_y_rsfmr_cor_gp_gp, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_rsfmr_cor_gp_gp}{Data file containing neurite density data}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
gord_cor Dataframe of white matter neurite densities
}
\description{
Extract cortical network correlations
}
