% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_headaches}
\alias{get_headaches}
\title{Get subject headache history}
\usage{
get_headaches(ph_p_mhx, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_mhx}{Data frame containing medical history}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing headache history
}
\description{
Get subject headache history
}
