% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_interview_age}
\alias{get_interview_age}
\title{Return dataframe containing interview age of specified subjects}
\usage{
get_interview_age(abcd_df, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_df}{Any ABCD dataframe containing interview age}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
interview_age Dataframe containing interview age
}
\description{
Return dataframe containing interview age of specified subjects
}
