% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_interview_age}
\alias{get_interview_age}
\title{return A data frame containing interview age of specified subjects}
\usage{
get_interview_age(abcd_y_lt, subjects = NULL, t = NULL, as_years = TRUE)
}
\arguments{
\item{abcd_y_lt}{Data frame containing interview age.}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{as_years}{Logical indicating whether to convert age to years.}
}
\value{
A data frame containing interview age.
}
\description{
return A data frame containing interview age of specified subjects
}
