% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_count}
\alias{get_mtbi_count}
\title{Extract number of mTBIs sustained by subject}
\usage{
get_mtbi_count(ph_p_otbi, abcd_y_lt, subjects = NULL, t = NULL, cutoff = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI data frame}

\item{abcd_y_lt}{Data frame containing age information}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{cutoff}{Maximum number of mtbis to be reported}
}
\value{
A data frame containing number of previous mTBIs
}
\description{
Extract number of mTBIs sustained by subject
}
