% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_loc}
\alias{get_mtbi_loc}
\title{Get acute symptom input variable 'latest_mtbi_loc'}
\usage{
get_mtbi_loc(ph_p_otbi, abcd_y_lt, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
mtbi_loc Dataframe containing latest_mtbi_loc
}
\description{
Get acute symptom input variable 'latest_mtbi_loc'
}
