% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_mechanism}
\alias{get_mtbi_mechanism}
\title{Get acute symptom input variable 'latest_mtbi_mechanism'}
\usage{
get_mtbi_mechanism(
  ph_p_otbi,
  abcd_y_lt,
  subjects = NULL,
  t = NULL,
  format = "undummied"
)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{subjects}{Vector of subjects to extract data for.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{format}{Variable indicating if df is dummied or undummied}
}
\value{
mtbi_mechanism Dataframe containing latest_mtbi_mechanism
}
\description{
Get acute symptom input variable 'latest_mtbi_mechanism'
}
