% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_mtbi_mem_daze}
\alias{get_mtbi_mem_daze}
\title{Get acute symptom input variable 'latest_mtbi_mem_daze'}
\usage{
get_mtbi_mem_daze(ph_p_otbi, abcd_y_lt, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI dataframe}

\item{abcd_y_lt}{Dataframe containing age information}

\item{subjects}{Vector of subjects to extract data for.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
mtbi_mem_daze Dataframe containing latest_mtbi_mem_daze
}
\description{
Get acute symptom input variable 'latest_mtbi_mem_daze'
}
