% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{get_mtbi_subjects}
\alias{get_mtbi_subjects}
\title{Extract mTBI subjects with a minimum time-since-last-mtbi threshold}
\usage{
get_mtbi_subjects(ph_p_otbi, abcd_y_lt, min_mpi = -10, t = NULL)
}
\arguments{
\item{ph_p_otbi}{TBI data frame}

\item{abcd_y_lt}{Data frame containing age information}

\item{min_mpi}{The minimum time-since-last-mtbi to be selected}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
subjects Data frame containing list of required subjects
}
\description{
Extract mTBI subjects with a minimum time-since-last-mtbi threshold
}
