% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_nihtbx_cardsort_fc}
\alias{get_nihtbx_cardsort_fc}
\title{Extract NIH toolbox cardsort data}
\usage{
get_nihtbx_cardsort_fc(abcd_tbss01, subjects = NULL, t = NULL)
}
\arguments{
\item{abcd_tbss01}{NDA nihtbx data frame}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing nihtbx cardsort data.
}
\description{
Extract NIH toolbox cardsort data
}
