% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_p_gender}
\alias{get_p_gender}
\title{Extract parent-reported gender of youth}
\usage{
get_p_gender(...)
}
\arguments{
\item{...}{Arguments passed to \code{get_gender_p()}.}
}
\value{
Data frame containing youth-reported gender.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for extracting parent-reported gender.
Please use \code{get_gender_p()} instead.
}
\keyword{internal}
