% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_prosocial_behaviour}
\alias{get_prosocial_behaviour}
\title{Extract prosocial behaviour}
\usage{
get_prosocial_behaviour(
  ce_p_psb,
  ce_y_psb,
  subjects = NULL,
  no_zero = FALSE,
  t = NULL
)
}
\arguments{
\item{ce_p_psb}{Parent Prosocial Behavior Survey}

\item{ce_y_psb}{Youth Prosocial Behavior Survey}

\item{subjects}{Vector of subjectkeys.}

\item{no_zero}{Boolean indicating if zero values should be replaced with 1}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
prosocial_behaviour
}
\description{
Extract prosocial behaviour
}
