% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_pubertal_status}
\alias{get_pubertal_status}
\title{Return dataframe containing pubertal status of specified subjects}
\usage{
get_pubertal_status(
  ph_p_pds,
  ph_y_pds,
  subjects = NULL,
  max_value = NULL,
  reports = "split",
  t = NULL
)
}
\arguments{
\item{ph_p_pds}{Dataframe containing parent pubertal status report}

\item{ph_y_pds}{Dataframe containing youth pubertal status report}

\item{subjects}{Vector of subjectkeys.}

\item{max_value}{Maximum value for pubertal status}

\item{reports}{If "split" (default), returns parent and youth PDS scores.
If "mean", returns the average of the parent and youth PDS scores.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
pubertal_status Dataframe containing average pubertal status
}
\description{
Return dataframe containing pubertal status of specified subjects
}
