% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_race}
\alias{get_race}
\title{Returns race information}
\usage{
get_race(
  abcd_p_demo,
  subjects = NULL,
  t = t,
  dummy = FALSE,
  asian_as_other = FALSE
)
}
\arguments{
\item{abcd_p_demo}{Dataframe containing parent demographic information}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}

\item{dummy}{String indicating format to output race data}

\item{asian_as_other}{If TRUE, Asian category is included in "other".}
}
\value{
race_df Dataframe containing subject race
}
\description{
Returns race data categorized by the ABCD Study based on the NIH Minimum
Reporting guidelines and the US Census Bureau's OMB standards.
If dummy = FALSE, five categorical variables for different race groups
are provided. Otherwise, a single categorical variable is returned.
}
