% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_sds_total_probs}
\alias{get_sds_total_probs}
\title{Deprecated function for pulling sleep disturbance scale total problems.}
\usage{
get_sds_total_probs(ph_p_sds, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_sds}{Dataframe containing sleep disturbance scale data.}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...).}
}
\value{
Data frame containing sleep data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Defunct function for converting a data list into a data frame. Please
use \code{as.data.frame()} instead.
}
\keyword{internal}
