% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_sports_and_activities}
\alias{get_sports_and_activities}
\title{Extract healthy behaviours: spots and activities questionnaire}
\usage{
get_sports_and_activities(
  ph_p_saiq,
  subjects = NULL,
  t = NULL,
  discretize = FALSE
)
}
\arguments{
\item{ph_p_saiq}{ABCD Parent Sports and Activities Involvement
Questionnaire}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}

\item{discretize}{Boolean indicating if measures should be discretized into
quartiles}
}
\value{
activities
}
\description{
Extract healthy behaviours: spots and activities questionnaire
}
