% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_sports_and_activities}
\alias{get_sports_and_activities}
\title{Extract healthy behaviours: spots and activities questionnaire}
\usage{
get_sports_and_activities(ph_p_saiq, subjects = NULL, t = NULL)
}
\arguments{
\item{ph_p_saiq}{ABCD Parent Sports and Activities Involvement
Questionnaire}

\item{subjects}{Dataframe containing list of required subjects}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ... 3: 3yfu)}
}
\value{
activities
}
\description{
Extract healthy behaviours: spots and activities questionnaire
}
