% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_subc_cor}
\alias{get_subc_cor}
\title{Extract subcortical network correlations}
\usage{
get_subc_cor(mri_y_rsfmr_cor_gp_aseg, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_rsfmr_cor_gp_aseg}{Data file containing neurite density data}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame of white matter neurite densities
}
\description{
Extract subcortical network correlations
}
