% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_subc_v}
\alias{get_subc_v}
\title{Extract subcortical volumes}
\usage{
get_subc_v(mri_y_smr_vol_aseg, subjects = NULL, t = NULL)
}
\arguments{
\item{mri_y_smr_vol_aseg}{Data file containing subcortical data}

\item{subjects}{Vector of subjectkeys.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
subc_v_df Dataframe of subcortical volumes
}
\description{
Extract subcortical volumes
}
