% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_y_gender}
\alias{get_y_gender}
\title{Extract youth-reported gender of self}
\usage{
get_y_gender(...)
}
\arguments{
\item{...}{Arguments passed to \code{get_gender_y()}.}
}
\value{
Data frame containing youth-reported gender.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated function for extracting youth-reported gender.
Please use \code{get_gender_y()} instead.
}
\keyword{internal}
