% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraction.R
\name{get_y_gender}
\alias{get_y_gender}
\title{return A data frame containing youth-reported genders}
\usage{
get_y_gender(gish_y_gi, subjects = NULL, t = NULL)
}
\arguments{
\item{gish_y_gi}{Youth-reported gender data frame}

\item{subjects}{Vector of subjects to filter to.}

\item{t}{timepoint of data collection (0: baseline, 1: 1yfu, ...)}
}
\value{
A data frame containing sex
}
\description{
return A data frame containing youth-reported genders
}
