% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{original_tbi_names}
\alias{original_tbi_names}
\title{Display conversion table of original and new TBI variable names}
\usage{
original_tbi_names()
}
\description{
This function neatly displays a table indicating the original and new
variable names for the TBI variables in the ABCD dataset following usage
of the \code{rename_tbi} function.
}
