% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_rsfmri}
\alias{qc_rsfmri}
\title{rsfMRI QC}
\usage{
qc_rsfmri(abcd_df, mriqcrp10301, t = 0, no_na = FALSE)
}
\arguments{
\item{abcd_df}{A dataframe containing subjectkeys}

\item{mriqcrp10301}{The \code{mriqcrp10301} data object}

\item{t}{The collection period of interest (defaults to baseline: 0)}

\item{no_na}{whether observations containing NAs should be removed}
}
\value{
rsfmri_qc_list a list of:
\enumerate{
\item The original dataframe with failing QC participants dropped
\item A list of subjectkeys of dropped QC participants
}
}
\description{
Given a raw MRI QC file and a dataframe containing subjects, removes subjects
that do not pass all rsfMRI quality control and protocol compliance checks
}
