% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_control.R
\name{qc_smri}
\alias{qc_smri}
\title{smri QC}
\usage{
qc_smri(
  abcd_df,
  mri_y_qc_raw_smr_t1,
  mri_y_qc_raw_smr_t2,
  t = 0,
  no_na = FALSE
)
}
\arguments{
\item{abcd_df}{A dataframe containing subjectkeys}

\item{mri_y_qc_raw_smr_t1}{QC info for T1 sMRI}

\item{mri_y_qc_raw_smr_t2}{QC info for T2 sMRI}

\item{t}{timepoint for data collection (0: baseline, 1: 1yfu, ...)}

\item{no_na}{whether observations containing NAs should be removed}
}
\value{
smri_qc_list a list of:
\enumerate{
\item The original dataframe with failing QC participants dropped
\item A list of subjectkeys of dropped QC participants
}
}
\description{
Given a raw MRI QC file and a dataframe containing subjects, removes subjects
that do not pass all smri quality control and protocol compliance checks
}
