% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbi.R
\name{rename_tbi}
\alias{rename_tbi}
\title{Rename non-descriptive columns in ABCD's TBI file.}
\usage{
rename_tbi(tbi_df)
}
\arguments{
\item{tbi_df}{The complete file abcd_otbi01.txt}
}
\value{
renamed_tbi A modified form of tbi_df with clearer column names
}
\description{
Renames column names of ABCD's Ohio TBI Screen data to meaningful names
that are easier to work with.
Use 'original_otbi_names()' to print out a conversion table of old and new
column names.
}
